+++ src/Tangle.Trading/Stock/Command/OrderSharesCommand.cs
@@ -13,7 +13,7 @@ namespace Tangle.Trading.Stock.Command
         public ORDER_SIDE Side { get; private set; }
         public int Quantity { get; private set; }
         public OrderType Type { get; private set; }
-        public float Price { get; private set; }
+        public decimal Price { get; private set; }
 
         /// <summary>
         /// 落指定股数的买/卖单，最常见的落单方式之一。可以下市价单或者限价单。
@@ -24,7 +24,7 @@ namespace Tangle.Trading.Stock.Command
         /// <param name="quantity">委托数量</param>
         /// <param name="type">委托方式</param>
         /// <param name="price">委托价格</param>
-        public OrderSharesCommand(string instrumentID, ORDER_SIDE side, int quantity, OrderType type, float price=0)
+        public OrderSharesCommand(string instrumentID, ORDER_SIDE side, int quantity, OrderType type, decimal price=0)
         {
             InstrumentID = instrumentID;
